Restructured cmake project
