+++ include/polygon_operations/utilities.h
