+++ test/convex_polygon_test.cpp
@@ -1,5 +1,5 @@
-#include "convex_polygon_library/convex_polygon.h"
-#include "convex_polygon_library/convex_hull.h"
+#include "polygon_operations/convex_polygon.h"
+#include "polygon_operations/convex_hull.h"
 #include "gtest/gtest.h"
 #include <random>
 #include <chrono>
